Combine RCS and old Github repo, preserving history
This new repo combines all the files from KA9Q's current tarball
with those that had been added to the old repo on Github.

Used the tool cvs-fast-export from ESR to translate the RCS history
into a format suitable for importing with git fast-import. Then
tried manually to reconcile the differences between the archives.

These files had newer checkins in RCS but with no changes other
than the Id: string. rcs-fast-export didn't include the trivially
changed checkins, and I left that decision alone:
  airspy.c audio.c audio.h control.h dump.c fm.h metadump.c
  modes.h pl.c touch.c

These files appeared in the RCS directory but were neither present
in HEAD on Github nor needed for a successful build, so they are
treated as deleted:
  c4fm.c cam.c command.h demod.c dependencies doc.txt dsb.c
  gentone.c gr.c hb15.s iq.c mirics.c multi_display.c nfuncube.c
  nmonitor.c notch.c pcm_monitor.c pmonitor.c rtp.h script
  send_status.c ssb.c tgen.c window.c window.h

This file appeared in RCS, and I moved it to a new doc/ folder:
  radio_multicast.numbers

These files appeared in the old Github repo, and I moved them
to the new doc/ folder:
  notes.txt hackrf-notes.txt dcc2018.pdf dcc2018.pages

I also added the RCS tarball to a new archive/ folder, just
in case the import process failed to preserve all the history
accurately.